Fix: Qualify only the helpers that can fold
`SZ_HELPER_AUTO` carries `constexpr` from C++20 onwards, which is what lets a CUDA
kernel reach a scalar helper without the C layer naming an execution space. A helper
that touches an intrinsic can never be constant-evaluated, though, and a `constexpr`
function with no constant-evaluated path is ill-formed: Clang and MSVC reject the
definition, GCC 12 too, and only GCC 13+ softens it to `-Winvalid-constexpr`. Carrying
the qualifier across every backend cost nine CI jobs at once - both GCC 12 runners,
both Windows runners under C3615, StringZillas-CPUs and StringZillas-CUDA whose host
compiler is `gcc-12`, and both Clang runners through the Rust `cc` build, which has no
warning flags of its own.

The 259 intrinsic-reaching helpers move to `SZ_HELPER_INLINE`, so no ISA backend header
holds a `SZ_HELPER_AUTO` at all, and `-Wno-invalid-constexpr` leaves the warning set
rather than spreading to `build.rs`: nothing that includes these headers needs a flag to
compile them. The function-local tables in seven backends become legal where they stand,
since only the qualifier made them a C++23 extension. The compiler enumerates the whole
set - `-Winvalid-constexpr` under GCC 13+, once per ISA - and its silence is the proof
the sweep is complete.

The four `substrings` helpers a CUDA walk calls keep the qualifier and now earn it: they
take `span<char const>` rather than punning a `span<byte_t const>` into the C iterators,
so nothing in the bodies blocks constant evaluation.

`szs_substrings_replace_*` spells its offsets `sz_size_t`, the width the engines address
the output tape with. Apple Clang had refused the mismatch outright, `sz_u64_t` being a
distinct type from `size_t` there rather than merely a same-width one.

+++ rust/stringzillas/substrings.rs
@@ -227,7 +227,7 @@ extern "C" {
         replacements: *const c_void, // SzSequence
         output_data: *mut u8,
         output_data_capacity: usize,
-        output_offsets: *mut u64,
+        output_offsets: *mut usize,
         output_bytes_written: *mut usize,
         error_message: *mut *const c_char,
     ) -> Status;
@@ -240,7 +240,7 @@ extern "C" {
         replacements: *const c_void, // SzSequence
         output_data: *mut u8,
         output_data_capacity: usize,
-        output_offsets: *mut u64,
+        output_offsets: *mut usize,
         output_bytes_written: *mut usize,
         error_message: *mut *const c_char,
     ) -> Status;
@@ -656,7 +656,7 @@ impl Substrings {
     ///
     /// Tape in, tape out: `overlap_policy` must name a non-overlapping cover, replacements are indexed
     /// by needle and inserted verbatim, and an empty replacement deletes. `output_offsets` takes
-    /// `haystacks.len() + 1` entries and is 64-bit so the pair reads directly as a `BytesTapeView<u64>`.
+    /// `haystacks.len() + 1` entries, in the width the engines address the output tape with.
     /// Returns the bytes written; size `output_data` with [`Substrings::replace_bound`] to be sure the
     /// call cannot be refused.
     pub fn replace_into<Replacement>(
@@ -666,7 +666,7 @@ impl Substrings {
         overlap_policy: OverlapPolicy,
         replacements: &[Replacement],
         output_data: &mut [u8],
-        output_offsets: &mut [u64],
+        output_offsets: &mut [usize],
     ) -> Result<usize, Error>
     where
         Replacement: AsRef<[u8]>,
@@ -972,7 +972,7 @@ mod tests {
         let bound = engine.replace_bound(replacements, input_bytes).unwrap();
 
         let mut output_data = vec![0u8; bound];
-        let mut output_offsets = vec![0u64; haystacks.len() + 1];
+        let mut output_offsets = vec![0usize; haystacks.len() + 1];
         let tape = AnyBytesTape::from_sequences(haystacks).unwrap();
         engine
             .replace_into(
@@ -987,8 +987,8 @@ mod tests {
 
         (0..haystacks.len())
             .map(|index| {
-                let start = output_offsets[index] as usize;
-                let end = output_offsets[index + 1] as usize;
+                let start = output_offsets[index];
+                let end = output_offsets[index + 1];
                 output_data[start..end].to_vec()
             })
             .collect()
@@ -1115,7 +1115,7 @@ mod tests {
         // Tape in, tape out: a rewrite has nowhere to put its product when the input is addressed by
         // callback, and the C API gives `replace` no sequence overload to try.
         let mut output_data = vec![0u8; 256];
-        let mut output_offsets = vec![0u64; corpus.len() + 1];
+        let mut output_offsets = vec![0usize; corpus.len() + 1];
         let refused = engine.replace_into(
             &device,
             &AnyBytesTape::from_slices(&corpus),
